fix: export the subscription options to avoid ts errors
